some code fixed, implemented array_size storage using size_t

+++ 3_arrays/include/array_utilities.h
@@ -8,9 +8,9 @@ typedef enum
     NULL_POINTER_ERROR
 } RemovalError;
 
-RemovalError remove_element_from_array(int* arr, int* arr_size, int element);
-RemovalError remove_element_from_array_preserve_order(int* arr, int* arr_size, int element);
-RemovalError remove_element_from_sorted_array(int* arr, int* arr_size, int element);
-RemovalError remove_element_from_sorted_array_preserve_order(int* arr, int* arr_size, int element);
+RemovalError remove_element_from_array(int* arr, size_t* arr_size, int element);
+RemovalError remove_element_from_array_preserve_order(int* arr, size_t* arr_size, int element);
+RemovalError remove_element_from_sorted_array(int* arr, size_t* arr_size, int element);
+RemovalError remove_element_from_sorted_array_preserve_order(int* arr, size_t* arr_size, int element);
 
 #endif
